adjust clipping plane equation for new panning feature
The original plane equation (A,B,C,D) was in world space and we need
all four values in model space.  The world origin transformed to model
space is no longer necessarily the model origin, now that we have
panning translations.

We can inverse-transform the world coordinate (0,0,0) and unit vector
endpoint (A,B,C) to get their model space equivalents. Their
difference is the new unit vector endpoint (A',B',C') which is
parallel to the original. But we also need to calculate a new plane
distance from model origin D' to finish the job, since the model
origin can be a different place entirely due to translation and
rotation.